+++ src/app/cart/item/item.component.ts
@@ -3,7 +3,8 @@ import {CartItem} from '../item.model';
 
 @Component({
     selector: 'app-cart-item',
-    templateUrl: './item.component.html'
+    templateUrl: './item.component.html',
+    styleUrls: ['./item.component.css']
 })
 
 export class ItemComponent {