+++ src/app/products/product/click-prevent.directive.ts
@@ -1,15 +1,17 @@
 import {Directive, EventEmitter, HostListener, Output} from '@angular/core';
 
 @Directive({
-    selector: '[appClickPrevent]'
+    // tslint:disable-next-line:directive-selector
+    selector: '[click.prevent]'
 })
 
 export class ClickPreventDirective {
-    @Output() appClickPrevent = new EventEmitter();
+    // tslint:disable-next-line:no-output-rename
+    @Output('click.prevent') clickPreventEvent = new EventEmitter();
 
     @HostListener('click', ['$event'])
     onClick(event) {
         event.preventDefault();
-        this.appClickPrevent.emit(event);
+        this.clickPreventEvent.emit(event);
     }
 }